requirements.txt
psycopg2-binary
beautifulsoup4
requests
pandas
rank_bm25 
sentence-transformers 
faiss-cpu
spacy
quantulum3
fastapi
uvicorn
streamlit
python-dotenv
pydantic
sqlite3
openai
